feat: Implement DryRunMapping RPC for testing mappings against sample JSON
- Move mapping engine from services/gateway/internal/mapping to
  shared/pkg/mapping so both gateway and reference-data can import it
- Gateway internal package re-exports engine via type aliases for
  backwards compatibility
- Add DryRunMapping RPC to MappingService proto with inbound/outbound
  direction, field-level trace, validation result and execution timing
- Add DryRunInbound/DryRunOutbound to shared Engine with per-field trace
  capture (source value, transformed value, transform type)
- Add GetByNameAndVersion to mapping.Repository for version-specific dry run
- Reference-data handler resolves latest ACTIVE when version omitted,
  or exact version when specified
