+++ shared/pkg/mapping/gjson.go
@@ -1,14 +1,3 @@
-// Package mapping provides utilities for transforming external JSON payloads
-// into internal Meridian gRPC request messages and vice versa.
-//
-// Field extraction from external payloads uses gjson path expressions,
-// which support dot notation, array indexing, and modifiers:
-//
-//	"customer.id"            — nested field
-//	"items.#.price"          — array element iteration
-//	"meta|@flatten"          — gjson modifier
-//
-// See https://github.com/tidwall/gjson for full path syntax documentation.
 package mapping
 
 import "github.com/tidwall/gjson"
